Project4/requirements.txt
Flask==2.3.2
Werkzeug>=3.0.3
tensorflow>=2.13.0
Pillow>=10.3.0
numpy>=1.24.0
opencv-python>=4.8.1.78
scikit-learn>=1.3.0
